+++ src/analysis/mod.rs
@@ -1,0 +1,7 @@
+//! Graph Analysis tools module:
+//! * Subgraphs
+//! * Graph partitioning
+//! * Motives spotting
+
+pub mod motives;
+pub mod subgraphs;
