+++ src/search/mod.rs
@@ -1,0 +1,6 @@
+//! Search tools module:
+//! * Taumode
+//! * Sorted Indexing
+
+pub mod sorted_index;
+pub mod taumode;
