+++ src/storage/tests/mod.rs
@@ -3,3 +3,6 @@ mod test_storage;
 
 #[cfg(test)]
 mod test_load_from_storage;
+
+#[cfg(test)]
+mod test_parquet;
